requirements.txt
streamlit==1.32.0
google-genai>=0.3.0
Pillow==10.3.0
python-dotenv==1.0.1
openai>=1.51.0
requests
protobuf